+++ src/main/java/database/storageEngine/page/PageFactory.java
@@ -1,4 +1,4 @@
-package database.innodb.page;
+package database.storageEngine.page;
 
 public class PageFactory {
