+++ src/main/java/database/storageEngine/page/PageManager.java
@@ -1,4 +1,4 @@
-package database.innodb.page;
+package database.storageEngine.page;
 
 import java.io.File;
 import java.io.FileInputStream;
