+++ src/main/java/database/storageEngine/page/StorageRecord.java
@@ -1,4 +1,4 @@
-package database.innodb.page;
+package database.storageEngine.page;
 
 import java.io.Serializable;
 import java.util.Arrays;
