+++ src/test/java/database/storageEngine/page/PageFactoryTest.java
@@ -1,4 +1,4 @@
-package database.innodb.page;
+package database.storageEngine.page;
 
 import static org.assertj.core.api.Assertions.assertThat;
